+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Re-add nodriver evaluate serialization options for improved JavaScript evaluation @dbrun3
+
 ### Added
 
 - Fix: allow reset expect and intercept @nathanfallet